+++ Interval.cpp
@@ -12,7 +12,7 @@ Interval::Interval(): m_min(+infinite), m_max(-infinite) {
 
 }
 
-Interval::Interval(const double min, const double max) : m_min(min), m_max(max) {
+Interval::Interval(const float min, const float max) : m_min(min), m_max(max) {
 
 }
 
@@ -23,50 +23,50 @@ Interval::Interval(const Interval &i1, const Interval &i2) {
 
 Interval &Interval::operator=(const Interval &interval) = default;
 
-Interval Interval::expand(double delta) const {
-    double pas = delta / 2;
+Interval Interval::expand(float delta) const {
+    float pas = delta / 2;
     return {m_min - pas, m_max + pas};
 }
 
-double Interval::min() const {
+float Interval::min() const {
     return m_min;
 }
 
-double Interval::max() const {
+float Interval::max() const {
     return m_max;
 }
 
-double Interval::size() const {
+float Interval::size() const {
     return m_max - m_min;
 }
 
-double Interval::center() const {
+float Interval::center() const {
     return m_min + (m_max - m_min) / 2;
 }
 
 Interval Interval::merge(const Interval &i1, const Interval &i2) {
-    const double min = i1.min() <= i2.min() ? i1.min() : i2.min();
-    const double max = i1.max() >= i2.max() ? i1.max() : i2.max();
+    const float min = i1.min() <= i2.min() ? i1.min() : i2.min();
+    const float max = i1.max() >= i2.max() ? i1.max() : i2.max();
     return {min, max};
 }
 
-void Interval::setMin(const double min) {
+void Interval::setMin(const float min) {
     m_min = min;
 }
 
-void Interval::setMax(const double max) {
+void Interval::setMax(const float max) {
     m_max = max;
 }
 
-bool Interval::contains(const double value) const {
+bool Interval::contains(const float value) const {
     return m_min <= value && value <= m_max;
 }
 
-bool Interval::surrounds(const double value) const {
+bool Interval::surrounds(const float value) const {
     return m_min < value && value < m_max;
 }
 
-double Interval::clamp(const double value) const {
+float Interval::clamp(const float value) const {
     if (value < m_min) return m_min;
     if (value > m_max) return m_max;
     return value;
