+++ Triangle.cpp
@@ -16,15 +16,15 @@ Triangle::Triangle(const Point3 &x, const Point3 &y, const Point3 &z, const Vec3
 
 bool Triangle::hit(const Ray &ray, Interval int_valid, Hit_record &rec) const {
     const Vec3 h = p_vect(ray.direction(), m_v);
-    const double a = p_scal(m_u, h);
+    const float a = p_scal(m_u, h);
     if (a == 0) return false;
 
     const Vec3 s = ray.origin() - m_origin;
-    const double u = p_scal(s, h)/a;
+    const float u = p_scal(s, h)/a;
     if (u < 0 or u > 1) return false;
 
     const Vec3 q = p_vect(s, m_u);
-    const double v = p_scal(ray.direction(), q) / a;
+    const float v = p_scal(ray.direction(), q) / a;
     if (v < 0 or v > 1 - u) return false;
 
     rec.m_t = -p_scal(p_vect(s, m_v), m_u) / a;
